requirements.txt
django==1.11.8 
djangorestframework==3.7.3
django-cors-headers==2.1.0
django-model-utils==3.0.0
psycopg2==2.7.3.2
docutils==0.14
python-dateutil==2.6.1
python-dotenv==0.8.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.10.0
urllib3==1.22
facebook-sdk==2.0.0